Bugs fixing

+++ Commands/CreateDirectoryCommand.cs
@@ -41,9 +41,13 @@ namespace FileBot.Commands
             await directoryRepository.Save();
 
             var markup = markupBuilder.Build(info.CurrentDirectory);
-
-            await client.DeleteMessageAsync(chatId, info.BotMessageId.Value);
-            Message message = await client.SendTextMessageAsync(chatId, info.CurrentDirectory.Name, replyMarkup: markup);
+            
+            if(info.BotMessageId is not null)
+            {
+                await client.DeleteMessageAsync(chatId, info.BotMessageId.Value);
+            }
+            
+            Message message = await client.SendTextMessageAsync(chatId, info.CurrentDirectory.Path, replyMarkup: markup);
 
             info.BotMessageId = message.MessageId;
             await userRepository.Save();
